feat: add request logging middleware
