+++ server/index.ts
@@ -3,66 +3,12 @@ import express from "express";
 import { registerRoutes } from "./routes";
 import { setupVite, serveStatic, log } from "./vite";
 import { errorHandler } from "./middlewares/error.middleware";
-
-const SENSITIVE_FIELDS = ["password", "email"];
-
-function maskSensitive(value: any): any {
-  if (Array.isArray(value)) return value.map(maskSensitive);
-  if (value && typeof value === "object") {
-    return Object.fromEntries(
-      Object.entries(value).map(([key, val]) =>
-        SENSITIVE_FIELDS.includes(key.toLowerCase())
-          ? [key, "***"]
-          : [key, maskSensitive(val)]
-      ),
-    );
-  }
-  return value;
-}
+import { loggingMiddleware } from "./middlewares/logging.middleware";
 
 const app = express();
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
-
-const detailedLogs =
-  process.env.DETAILED_LOGS === "true" || process.env.DETAILED_LOGS === "1";
-
-app.use((req, res, next) => {
-  const start = Date.now();
-  const path = req.path;
-  let capturedJsonResponse: Record<string, any> | undefined = undefined;
-
-  if (detailedLogs) {
-    const originalResJson = res.json;
-    res.json = function (bodyJson, ...args) {
-      capturedJsonResponse = bodyJson;
-      return originalResJson.apply(res, [bodyJson, ...args]);
-    };
-  }
-
-  res.on("finish", () => {
-    const duration = Date.now() - start;
-    if (path.startsWith("/api")) {
-      let logLine = `${res.statusCode} ${req.method} ${path} in ${duration}ms`;
-
-      if (detailedLogs) {
-        const payload = maskSensitive({
-          request: req.body,
-          response: capturedJsonResponse,
-        });
-        logLine += ` :: ${JSON.stringify(payload)}`;
-      }
-
-      if (logLine.length > 80) {
-        logLine = logLine.slice(0, 79) + "…";
-      }
-
-      log(logLine);
-    }
-  });
-
-  next();
-});
+app.use(loggingMiddleware);
 
 (async () => {
   const server = await registerRoutes(app);
